twister: move testsuite, testcase classes out
Move testsuite/testcase and supporting classes into own files.
