+++ scripts/pylib/twister/mixins.py
@@ -1,0 +1,8 @@
+# vim: set syntax=python ts=4 :
+#
+# Copyright (c) 2018 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+class DisablePyTestCollectionMixin(object):
+    __test__ = False
+
